Moved files to new folders and added sys.path where needed. TODO, fix win os error per issue 86
