+++ netcode/settings.py
@@ -1,0 +1,15 @@
+#Author:    John P Armentor
+#email:     johnparmentor@gmail.com
+#Date:      2020 01 30
+#Course:    CSC424 - Software Engineering II
+#Prof:      Dr. A. Louise Perkins
+#
+#File for designating settings intended across the entire game.
+
+import socket  
+
+# Multiplayer settings
+#
+MYIP = socket.gethostbyname(socket.gethostname())
+
+LOCALHOST = "192.168.1.64"
