+++ sandbox.py
@@ -9,6 +9,7 @@ sys.path.append('./player_character/skills')
 sys.path.append('./game_items')
 sys.path.append('./game_engine')
 sys.path.append('./story_items')
+sys.path.append('./netcode')
 
 import player_character
 import character_creation
