+++ server.py
@@ -3,7 +3,7 @@
 #Date:      2020 01 30
 #Modified:      2020 02 10
 #Course:    CSC424 - Software Engineering II
-#Modified:      2020 03 04
+#Modified:      2020 03 05
 #Course:    CSC425 - Software Engineering II
 #Prof:      Dr. A. Louise Perkins
 
@@ -16,6 +16,7 @@ sys.path.append('./player_character/abilities')
 sys.path.append('./player_character/skills')
 sys.path.append('./game_items')
 sys.path.append('./game_engine')
+sys.path.append('./netcode')
 
 import player_character
 import character_creation
